% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_mean.R
\name{geom_mean}
\alias{geom_mean}
\title{Calculate geometric mean of a numeric vector}
\usage{
geom_mean(x)
}
\arguments{
\item{x}{numeric vector of length >= 1.}
}
\value{
Geometric mean of elements in x.
}
\description{
Calculate geometric mean of a numeric vector
}
\examples{
geom_mean(1:10)
}
